Generate coverage report from all test runs
We're currently only generating a coverage report from the cover
target. This means that we get coverage only from the basic test
run without Praefect or WAL enabled. This commit changes the test-go
target to generate coverage report so the other jobs that build on it
will also have a coverage report generated. While the report is
generated, the coverage information is not printed by default. cover
target is left in place to print the coverage information after a
test run.