Gather coverage from all go test jobs
We're currently gathering coverage only from the test:coverage job.
This doesn't run with Praefect nor the WAL enabled. Since the test-go
target is now generating a coverage report for all of the test runs,
let's upload the report from each job to get a more complete picture
of the test coverage.

The test:coverage job is still left in place as it parses the total
test coverage percentage through the regex.