datastore: Convert pgx library to v5
Upgrade the pgx library to v5. The most notable change that we need to
adapt to is that the old "prefer_simple_protocol=true" connection option
has been reworked to "default_query_exec_mode".

Other than that, the major release has also introduced support for
generics. Unfortunately, we cannot really use them given that we use the
`database/sql` interface instead of the native pgx one. Ideally, we'd
thus migrate to use the native interface, but we can't easily do that
because the `rubenv/sql-migrate` library depends on the `database/sql`
interface directly.

Note that this does not yet migrate the pgtypes dependency. This will be
done in the next commit.