glsql: Upgrade the pgtype dependency to v5
The pgtype dependency has received a major upgrade to v5. This release
brings a dramatically refactored API that makes more use of generics in
many places. Unfortunately, we can't really use them as they do not work
with the generic `database/sql` interfaces, which we're using.

Our `StringArray` helper type depends on `pgtype.TextArray`, which has
been removed in favor of generics. Adapt the type accordingly.