ci: Add manual job to build proto gem
Because the Gitaly proto gem is only created and published on a git tag,
it's hard to test this. This change adds a manual job to build the proto
gem in CI. Although this does not test the publishing of the gem,
because we don't want to litter rubygems.org with tests, it exposes the
output gem as an artifact so it can be downloaded of inspection and
local testing.