ci: Generate code coverage for all test jobs
We have an explicit "coverage" job that computes test coverage. Now that
we have started to compute code coverage in our tests by default though
it doesn't make much sense anymore to treat this job specially.

Refactor our CI pipeline to extract code coverage for all jobs. This
gives us a better global picture of which jobs cover what parts of our
codebase. Remove the Makefile target that drove this as it is likely no
longer of use to anybody.