ci: Start testing with Go 1.19 and 1.20
We're about to bump the minimum required Go version to 1.19, so this
commit drops test jobs which exercise Go 1.18 and adds test jobs for the
new Go 1.20 release.

Note that we're not yet introducing tests with Go 1.20 for our FIPS test
jobs given that there is no FIPS-enabled image for Go 1.20 yet.