Assert the actual contents of HEAD
TransactionManager's tests are currently asserting the state of HEAD
through GetDefaultBranch. This includes also some fallback logic and
doesn't always return the actual value of HEAD. When we're testing
that the disk state matches what we expect, we care about the actual
state and don't want the assertions to be mangled by the legacy fallback
logic. This commit changes the TransactionManager tests to assert the
actual contents of HEAD.