praefect: Remove old sub-command implementation leftovers
All sub-commands of Praefect were moved to a third party
urfave/cli/v2 library. The old helper functions and an interface
are removed as they are not used anymore.
Also, the help text generated by new library now contains information
about all sub-commands and doesn't need to be modified.
The 'serve' sub-command can now be restricted to disallow arguments
to be passed for it.

Part of: #5001