praefect: Drop feature flag that fixes routing with additional repos
When rewriting the target repository, the rewritten relative path will
be specific to the node's storage to which the rewritten request is
being routed to. Consequentially, if we're also rewriting an additional
repository, it follows that both repositories must be hosted by the same
storage, or otherwise one of both repositories wouldn't actually be
resolveable on the local storage.

This property wasn't followed when routing repository creations that
have an additional repository. This can lead to a case where we resolve
the additional repository first, but ultimately assign the newly created
repository to nodes that don't actually have the additional repository.
The result is breakage.

We have fixed this bug in 41c10e01f (praefect/router: Fix creation of
repos with additional referenced repo, 2023-05-10), where we started to
route repository creating RPCs to the same nodes as the additional
repository is assigned to. This change was rolled out more than a week
ago into production systems without any observed issues.

Remove the feature flag guarding the code so that we unconditionally
route such repository creating requests to the correct node.