housekeeping: Fix race condition in `addPackRefsInhibitor`
The `addPackRefsInhibitor` function allows users to add an inhibitor to
`git-pack-refs(1)`. In the function we check if `git-pack-refs(1)` is
currently/still running (in a loop) and allow the user to exit early via
context cancellation.

There is race condition here wherein we read from the
`state.packRefsDone` channel without a lock. This is a race condition,
since, the cleanup function of `tryRunningPackRefs` could also set
`state.packRefsDone` to nil concurrently. And reading from a nil channel
causes a panic.

We solve this race by using the local variable `packRefsDone`, which is
set after obtaining a lock. This way the local variable could be closed
but will not be `nil`, since only `state.packRefsDone` is set to `nil`.