objectpool: Make alternates disconnection reusable
The `DisconnectGitAlternates()` RPC handles disconnecting a repository
from an object pool. In the future, the `ReplicateRepository()` RPC will
need to also be capable of disconnecting a target repository from its
object pool when the source repository is not linked. This change
factors the object pool disconnection logic out of the RPC and moves it
into the `git/objectpool` package.