Log a directory instead of a single pack file
Each transaction has a staging directory where they can stage files
for commit. Currently it is used to store two things.

- stagingDir/quarantine stores the transaction's quarantine directory.
  This is where all the RPC handlers write the git objects prior while
  staging the transaction prior to a commit.
- stagingdir/transaction.pack is created right before committing the
  transaction to create a pack file that can be logged.

The current assumption is that we're only ever logging a single file,
which is the pack file containing the transaction's objects. However,
there are other files we could want to log along the pack file. For
example, we could compute and log the pack file's index to avoid having
to do so when applying the log entry.

This commit changes the transaction.pack to be placed in
stagingDir/wal-files/transaction.pack. Instead of logging only the pack
file, we log the entire wal-files directory. This makes room for logging
other files alongside the pack file as well.

There's no changes in external behavior.