Provide path to parser functions in RequireDirectoryState
Depending on the file in question, one might want to do assertions
not only based on the file's content but input. This is the case
for example when asserting pack file indexes are correct. We don't
really care about the content of the actual file but we want to make
sure it's a valid index for a given pack. Provide the path in the
ParseContent method to make it easier to distinguish which file is
being parsed.