Apply logged pack files to repository without copying
TransactionManager is currently unpacking pack files in order to
apply them to the repository from the log. This is inefficient:

1. The objects are copied from the pack file when they are unpacked.
2. Accessing unpacked objects is less efficient than packed ones.
3. It's more work to repack the objects later.

To avoid these problems, this commit instead hard links the logged
pack files from the log to the repository. This is done as follows:

1. When staging the transaction, a pack file is computed that contains
   objects that are newly reachable from the new reference tips. An
   index is also computed for the pack file.
2. The pack file and its index are logged.
3. The pack file and the index are hard linked from the log directory
   to the repository's 'objects/pack' directory.

The pack file and the index are logged with the name Git gives them, so
`pack-<digest>.{pack,idx}`. It would be simpler to log the files with
static names so they are always named the same for all transactions,
say `transaction.{pack,idx}`. They could then be linked to the repository's
object directory under the log entry's index, so for example
`objects/pack/<log_index>.{pack,idx}`. This would be simpler as we wouldn't
have to pipe the pack prefix through the log. The problem is that Git
doesn't seem to automatically remove these packs when it's doing a
full repack with `git repack -ad`. This would lead to the packs
accumulating in the repository. For that reason, we use the Git generated
names.

Applying pack files directly doesn't also come without downsides.
With each write resulting in an additional pack file in the repository,
looking up objects from the packs becomes slower. This is particularly
a problem if there are large number of small writes. We might want
to later have a look at using a threshold on the object count in the
pack to decide whether we apply it directly or unpack it similarly to
what Git does when receiving packs. We'll leave this for later though
once we know this is a problem.