Always initialize staging directory for a transaction
So far we've been initializing the staging directory only when a
transaction is about to write objects. Since we're soon about to
set up a quarantine automatically when the transaction begins,
we'll need the staging directory to store it. Let's create the
staging directory thus unconditionally.

Tests had to be adapted as the assertions ensuring the staging
directory is always cleaned up were failing. Some of the tests were
not always finishing their transactions. This was fine as they didn't
write objects so the staging directory was not created and there was
no state to clean up. Now that we always create it, we need to finish
the transactions so they clean up after themselves.

If there is a crash, some transactions could be interrupted without
being able to clean up after themselves. This is handled in the
PartitionManager by always deleting and recreating the staging directory
on partition start to ensure it doesn't contain any stale data. Do
the same in the TransactionManager tests to clean up after crashes.

Previously read-only transactions would not set up a quarantine to avoid
wasteful work that is not needed. They'll do so for now but we'll later
fix this by differentiating read-only transactions from read-write
transactions.