Setup quarantine directory unconditionally
Transaction is currently setting up the quarantine directory once
the QuarantineDirectory method is called. Integrating this across
the whole application code is a lot of work. It would be better if
we could stay behind the existing APIs to make the integration easy.
To do so, we'll soon support rewriting the repository from the
request to correctly wire it up with the quarantine. With that, we'll
be able to wire up the quarantine in automatically. To facilitate this,
let's create the quarantine always when beginning a transaction.

With the quarantine directory always created, we also remove the
QuarantineDirectory method. It was previously used by the tests to
determine where the objects written into the quarantine should go.
For now, we take the quarantine directory directly from the transaction
in the tests. We'll later make use of the new interface that returns
the repository details.

The packObjects method was previously using the existence of the
quarantine directory as a marker that something has been written into
it. Since we no longer have the method, we'll do the same check more
generally by looking into the directory to see if there's new content
there or not.