Unconditionally setup a staging repository
We're currently setting up the staging repository when we are about
to commit a transaction. Staging repository is essentially the target
repository with the transaction's quarantine directory applied. We're
looking to automatically quarantine the repositroy for the RPC handlers
and returning the details how to access a repository from the transaction
to make integration easier. To facilitate that, setup the staging
repository already on transaction beginning. We can then use it later
rewrite the request's repository.