Implement RewriteRepository on Transaction
We were previously exposing the QuarantineDirectory() directly from
the Transaction. The idea was that we'd wire the RPC handlers to use
the quarantine directory for their Git invocations. This would have
been a lot of work as we'd have to touch all RPCs separately that
were setting up a quarantine directory. This commit introduces the
RewriteRepository() method which instead rewrites the provided repository
to correctly access the repository in the scope of the transcation.
This will later be used to automatically wire repositories correctly.