Use localrepo.UnpackObjects for writing to quarantine
The TransactionManager tests are currently testing the quarantining
logic by writing pack files in to the quarantine directory of the
Transaction. Doing so they manually resolve the quarantine directory
path from the *gitalypb.Repository. Write the objects into the
quarantine by unpacking them using localrepo.UnpackObjects. This better
tests that the repository is correctly wired up with a quarantine in
a manner our git helpers handle them.

In the tests, the objects were previously in the quarantine as packs.
Now they are unpacked into the quarantine. This makes no difference
when it comes to external behavior though.