git2go: Use struct slice instead of map in `conflicts`
The `Merge()` function iterates over a map and does a lookup on blobs.
This works, but iterating over a map doesn't guarantee the order of
iteration. Matter of fact, go randomizes map iteration to avoid sloppy
programming.

The iteration order doesn't really matter for the functionality of the
function. The only problem is it makes testing non-deterministic. Since
we're about to rid this code with time and to make the upcoming tests
deterministic, lets replace the map with a struct slice.