ref service: Use HeadReference instead of GetDefaultBranch
FindAllBranches was using GetDefaultBranch when checking for merged
branches, but this method has some unintuitive behaviour. Namely that it
wont use HEAD if the repository is empty and it will select any branch
if that's the only branch that exists.

Neither of those edge cases make sense for a branch listing.