repository: Move GitRepo interface into storage package
The `repository.GitRepo` interface provides a set of functions that
allow us to locate a repository in our storage. It lives in a separate
package outside of the `internal/git` module in order to avoid a set of
cyclic dependencies. The current location is kind of weird though, as
this interface is inherently tied to our storage details instead of
being a generic Git thing.

Move the definition of the interface into `internal/gitaly/storage` so
that it is defined close to where it is used. This should also be a
relatively safe location as any package that depends on the interface
should implicitly already depend on the storage package given that it
will typically be passed down into any kind of storage locator.

While at it, rename it to `storage.Repository`.