praefectutil: Move functionality relating to storage paths
The praefectutil package contains a set of functions that compute and
classify repository paths for Praefect's rewritten replica paths. They
had to be split out into a separate package given that they're also used
by Gitaly code in order to figure out whether a repository is an object
pool or not. Furthermore, the code to figure out whether a repository is
an object pool or not is also partially contained in the stats package,
which adds even more to the confusion as it is now split up across
multiple packages.

Move the computations for Praefect paths into our storage package. This
makes for a better fit as this package should be the only source of
truth for where repositories are located anyway, and is already used by
both Praefect and Gitaly.

The next commit will also move the code from the stats package in here
to make the knowledge neatly contained in a single place.