git: Convert `ValidateRevision()` to accept options
We're about to extend `ValidateRevision()` so that it can optionally
handle pseudo-revisions like `--all` and `--not` as well as path-scoped
revisions. As not all existing callers will want to make use of these we
need to allow callers to opt-in to this new behaviour.

Merge `ValidateRevision()` as well as `ValidateRevisionAllowEmpty()`
into a single function that can be passed an option. As a first example,
we also introduce the `AllowEmptyRevision()` option that gets passed
from all callers which were calling `ValidateRevisionAllowEmpty()`.