blob: Convert `ListLFSPointers()` to use `VerifyRevisions()`
The `ListLFSPointers()` RPC has an open-coded version of revision
validation. This validation is both too lenient and too restrictive as
it does not reject known-bad patterns while also restricting
pseudo-revisions which would be perfectly fine.

Refactor the RPC to use `ValidateRevision()` as our new single source of
truth for revision validation.