git: Optionally allow pseudo-revisions in `ValidateRevision()`
As a safety guard against command-line options injection we disallow
passing any revisions that have a dash as leading character. There are
valid usecases where a caller may want to pass these though, namely in
the form of pseudo-revisions like `--all`, `--not` or `--glob=`.

Add a new option `AllowPseudoRevisions()` that puts known
pseudo-revisions onto an allowed-list.