repository: Disable `fsck` during replication
The `ReplicateRepository()` RPC uses `git-fetch(1)` to retrieve objects
from preexisting internal repositories. Currently the `fetch` also
performs consistency checks on the newly retrieved objects. This is
problematic because preexisting objects should alway be able to be
replicated regardless of the consistency check. This change sets
`fetch.fsckObjects` to `false` for the performed `fetch` to disable
`fsck` consistency checks.