storage: Consolidate `GetPath()` in favor of `GetRepoPath()`
The `storage.Locator` interface provides two functions `GetPath()` and
`GetRepoPath()` which both retrieve the path of a repository. It isn't
exactly easy to tell which one does what though. To shed some light on
it: the only difference is that `GetPath()` returns the repository path
without verifying that the repository is valid, while `GetRepoPath()`
does verify the repository location.

To improve upon this confusing interface we have introduced options for
`GetRepoPath()` that allow the caller to skip verification in 854b62c45
(storage: Introduce `GetRepoPathConfig`, 2023-03-27). Let's thus perform
the migration and use this function exclusively now, dropping the old
`GetPath()` function.