commonerr: Move repository-related errors into storage package
The commonerr package bundles a set of common errors that we use across
the project. Right now the current set only includes storage-related
errors though that relate to repositories being missing or invalid. And
while we use these common errors only in the Praefect codebase right
now, it would make sense to have Praefect and Gitaly share these errors
so that we ideally have a single source of truth for these.

Move the errors into the storage package so that it becomes less awkward
to share them between Gitaly and Praefect. This feels like a natural
place for these errors given that the errors can be directly attributed
to the storage level.