storage: Provide information about why repositories are invalid
The `IsGitDirectory()` function accepts a path and then checks whether
it corresponds to a valid Git repository or not. This information is
returned via a boolean, but that is indeed quite lacking given that the
caller cannot figure out _why_ a specific path would not be a valid
repository. This is noticable in our error reporting, where we don't
discern between a missing and a corrupt repository.

Refactor the function to instead return an error, renaming it to
`ValidateRepository()` to better match its new semantics. The newly
introduced errors are not yet being used to improve our error reporting
as it would be seen in our logs.