proto: Add toggle for object pool replication
The `ReplicateRepository()` RPC is not able to replicate object pool
relationships on source repository to the target repository. Currently
when repository replication occurs, the target repository fetches from
the source repository and ignores object pool relationships.

In the future, `ReplicateRepository()` will need the ability to recreate
object pool relationships that exist in the source repository on the
target repository. This behavior should be optionlly available in order
to preserve the existing behavior for clients that depend on it.

The `replicate_object_deduplication_network_membership` field is
introduced as part of the `ReplicateRepository()` request to enable
object pool replication. At this point it serves as a place holder for
the object pool relication feature.