ci: Use ramdisk to speed up tests on macOS
The macOS-based test jobs are extremely slow, typically taking about 55
minutes or more. It seems like the root cause for this is the ton of I/O
that we perform during our tests, where the disk becomes our bottleneck.

Convert the job to use a RAM disk for our test data. This results in a
significant speedup of test execution from around 2500 seconds to only
around 600 seconds.