Makefile: Make generation of coverage info optional
Starting with 77c67bb94 (Generate coverage report from all test runs,
2023-05-21), we generate coverage information by default. The runtime
overhead of recording the data is comparatively negligible with a 3%
slowdown. But interestingly, it does take quite a while to generate the
resulting Coverage report.

Fix this by making generation of coverage info optional with the newly
introduced `TEST_WITH_COVERAGE` variable. This allows us to only compute
the data as needed and provides a very noticable speedup:

    Benchmark 1: make test-go
      Time (mean ± σ):     69.015 s ±  1.001 s    [User: 521.524 s, System: 171.512 s]
      Range (min … max):   67.865 s … 69.690 s    3 runs

    Benchmark 2: make test-go TEST_WITH_COVERAGE=YesPlease
      Time (mean ± σ):     120.624 s ±  1.185 s    [User: 673.246 s, System: 207.668 s]
      Range (min … max):   119.470 s … 121.838 s    3 runs

    Summary
      'make test-go' ran
        1.75 ± 0.03 times faster than 'make test-go TEST_WITH_COVERAGE=YesPlease'

As we do indeed want to have coverage information available in our CI
systems the variable is set in our test jobs by default.