praefect: Properly set up logging interceptors
The Gitaly server is setting up a message producer that provides a bunch
of functionality:

    - It provides information about how many bytes have been received
      and sent to the client.

    - It allows filtering of gRPC log messages by full method via
      environment variables.

    - It injects metadata of structured errors into the log message.

Praefect doesn't yet set up this message producer and is thus lacking
all of these features.

Set up the message producer properly. While we only really care about
injection of structured errors in this patch series, it does not hurt to
also provide the other features while we're already at it.