praefect: Introduce options for server construction
In order to easily test structured errors we need to add interceptors
defined in the testserver package. As this is only done for testing
purposes we need the ability to change the way the server's interceptors
are set up.

Add the ability to pass server options into `NewGRPCServer()` as well as
the `NewServerFactory()` functions and implement options to add unary as
well as stream interceptors.

Note that this replaces the previous `grpc.ServerOption` parameter,
which wasn't used anywhere. It can be readded at a later point via a
separate option if required.